refactor ui trace list sections
